agent/pyproject.toml
[project]
name = "aura-chef-agent"
version = "1.0.0"
description = "A LangGraph agent served via FastAPI with CopilotKit integration"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langgraph>=1.0.5",
    "langchain-google-genai>=2.0.5",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.5,<1.0.0",
    "uvicorn>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "copilotkit>=0.1.74",
    "ag-ui-langgraph>=0.0.20",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "instaloader>=4.13.0",
    "yt-dlp>=2024.0.0",
    "lxml>=5.0.0",
    "opik>=1.0.0",
]
